+++ src/mail/mail.controller.ts
@@ -4,11 +4,11 @@ import { IContactFormAdminContext } from './interfaces/contact-form-admin-contex
 import { IContactFormUserContext } from './interfaces/contact-form-user-context.interface';
 import { IUpdateOrderContext } from './interfaces/update-order-context.interface';
 import { ApiOperation, ApiTags, ApiBody } from '@nestjs/swagger';
-import { ContactFormDto, UpdateOrderDto } from './dto/sendMail.dto';
+import { ContactFormDto } from './dto/contactForm.dto';
+import { UpdateOrderDto } from './dto/updateOrder.dto';
 import { Cases } from 'src/mail/dto/sendMail.dto';
 import { ADMIN_EMAIL } from 'src/config/env';
 import { ConsultationReason } from './interfaces/update-order-context.interface';
-import { HttpException } from '@nestjs/common/exceptions';
 
 
 @ApiTags('Mail')
@@ -26,28 +26,26 @@ export class ContactController {
     try {
       // Obtener el motivo de la consulta
       const consultationReason: ConsultationReason = updateOrderData.consultationReason;
-
-      // Contexto para el administrador
-      const updateContext = {
+      const orderContext: IUpdateOrderContext = {
         name: updateOrderData.name,
         phone: updateOrderData.phone,
         message: updateOrderData.message,
         userEmail: updateOrderData.userEmail,
         order: updateOrderData.order,
         consultationReason: updateOrderData.consultationReason,
-      };
-
-      // Contexto para el usuario
-      const userContext = {
-        firstname: updateOrderData.name,
-      };
+      }
 
       // Enviar correos electrónicos
       await this.mailService.sendMails({
         addressee: ADMIN_EMAIL,
         subject: Cases.UPDATE_ORDER,
-        context: updateContext,
-      });
+        context: orderContext,
+      });      
+
+      // Contexto para el usuario
+      const userContext = {
+        firstname: updateOrderData.name,
+      };
 
       await this.mailService.sendMails({
         addressee: updateOrderData.userEmail,
@@ -57,16 +55,13 @@ export class ContactController {
 
       return {
         status: 200,
-        message: 'Solicitud de cambio en la orden enviada con éxito',
-      };
+        message: 'Solicitud de cambio en la orden enviada con exito',
+      }
     } catch (error) {
-      throw new HttpException({ 
-        status: error.status, 
-        error: error.message, 
-        cause: new Error("Some Error")
-      }, error.status);
+      throw new InternalServerErrorException(error)
     }
   }
+
 
   @ApiOperation({
     summary: 'Ruta para contact form del front',
